Add arguments and pretty colored output to upgrade-scripts.sh

+++ upgrade-scripts.sh
@@ -2,43 +2,61 @@
 
 set -euo pipefail
 
-function errcho {
-    >&2 echo "$@"
-}
-
-function recreate_dir {
-    dir=$1
-    mkdir -p "${dir}"
-    ls -A1 "${dir}" | while read file; do
-        [ -z "${file}" ] && continue
-        rm -rf "${dir}/${file}"
-    done
-}
+source "scripts/internal/util.sh"
 
 function usage {
-	errcho "Usage: upgrade.sh <problem-dir>"
-	exit 2
+	errcho "Usage: upgrade.sh [options] <problem-dir>"
+	errcho "Options:"
+	errcho -e "  -h, --help"
+	errcho -e "  -d, --dry"
 }
 
-[ $# -eq 1 ] || usage
 
-problem_dir="$1"
-problem_scripts="${problem_dir}/scripts"
-source_current_scripts="scripts"
-old_scripts="old"
-problem_old_scripts_version_file="${problem_dir}/.scripts_version"
 dry_run="false"
 
+function handle_option {
+    shifts=0
+    case "${curr}" in
+        -h|--help)
+            usage
+            exit 0
+            ;;
+        -d|--dry)
+            dry_run="true"
+            ;;
+        *)
+            invalid_arg "undefined option"
+            ;;
+    esac
+}
 
-if [ -z "${problem_dir}" ]; then
-    errcho "Problem directory is not specified."
+function handle_positional_arg {
+    if [ -z "${problem_dir+x}" ]; then
+        problem_dir="${curr}"
+        return
+    fi
+    invalid_arg "meaningless argument"
+}
+
+argument_parser "handle_positional_arg" "handle_option" "$@"
+
+
+if [ -z "${problem_dir+x}" ]; then
+    cecho red >&2 "Problem directory is not specified."
+    usage
     exit 2
 fi
 
 if [ ! -d "${problem_dir}" ]; then
-    errcho "Problem directory '${problem_dir}' not found"
+    cecho red >&2 "Problem directory '${problem_dir}' not found"
     exit 4
 fi
+
+
+problem_scripts="${problem_dir}/scripts"
+source_scripts="scripts"
+old_scripts="old"
+problem_old_scripts_version_file="${problem_dir}/.scripts_version"
 
 
 if [ -f "${problem_old_scripts_version_file}" ]; then
@@ -48,7 +66,7 @@ else
 fi
 
 if ! git show -s --oneline "${old_scripts_version}" > /dev/null; then
-    errcho "Problem scripts version '${old_scripts_version}' not found in this source repo"
+    cecho red >&2 "Problem scripts version '${old_scripts_version}' not found in this source repo"
     exit 4
 fi
 
@@ -67,15 +85,15 @@ function source_current_scripts_version {
 function list_dir_files {
     dir="$1"
     pushd "${dir}" > /dev/null 2>&1
-    find . -type f
+    git ls-files
     popd  > /dev/null 2>&1
 }
 
 function get_all_scripts_files {
     {
-        list_dir_files "${source_current_scripts}"
+        list_dir_files "${source_scripts}"
         list_dir_files "${problem_scripts}"
-    } | cut -d/ -f2- | sort | uniq
+    } | sort | uniq
 }
 
 has_conflicts="false"
@@ -91,7 +109,13 @@ function push_conflict {
 
 function do_action {
     if ! "${dry_run}"; then
-        "$@"
+        status="FAIL"
+        ret=0
+        "$@" || ret=$?
+        if [ ${ret} -eq 0 ]; then
+            status="OK"
+        fi
+        return ${ret}
     fi
 }
 
@@ -99,11 +123,12 @@ recreate_dir "${old_scripts}"
 
 get_all_scripts_files | while read file; do
     c="${problem_scripts}/${file}"
-    b="${source_current_scripts}/${file}"
+    b="${source_scripts}/${file}"
     a="${old_scripts}/${file}"
 
     a_exists="true"
-    git show "${old_scripts_version}:${file}" > "${a}" || a_exists="false"
+    mkdir -p "$(dirname "${a}")"
+    git show "${old_scripts_version}:${source_scripts}/${file}" > "${a}" 2> /dev/null || a_exists="false"
 
     b_exists="true"
     [ -f "${b}" ] || b_exists="false"
@@ -112,6 +137,7 @@ get_all_scripts_files | while read file; do
     [ -f "${c}" ] || c_exists="false"
 
     status="OK"
+    "${dry_run}" && status="SKIP"
     message=""
     if "${b_exists}"; then
         if "${c_exists}"; then
@@ -150,25 +176,32 @@ get_all_scripts_files | while read file; do
                 message="CONFLICT! file '${file}' which is changed in problem has been deleted"
             fi
         else
-            errcho "Unreachable code reached!"
+            cecho red >&2 "Unreachable code reached!"
             exit 1
         fi
     fi
 
-    errcho -e "${file}:\t[${status}]\t${message}"
+    printf >&2 "%-40s" "${file}"
+    box_padding=10
+    echo_status "${status}"
+    errcho "${message}"
 done
 
-echo "$(source_current_scripts_version)" > "${problem_old_scripts_version_file}"
+if ! "${dry_run}"; then
+    echo "$(source_current_scripts_version)" > "${problem_old_scripts_version_file}"
+fi
 
-echo
+errcho
 
 
 if "${has_conflicts}"; then
-    errcho "Please resolve the following conflicts now!"
+    cecho red >&2 "Please resolve the following conflicts now!"
     sleep 0.1
     for file in "${conflict_list[@]}"; do
         echo "${file}"
     done
 
     exit 10
+else
+    cecho green >&2 "OK"
 fi
